# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School CRM API Documentation",
    "version": "1.0.0",
    "description": "Complete API documentation for School CRM SaaS Platform",
    "contact": {
      "name": "API Support",
      "email": "support@schoolcrm.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.schoolcrm.com/api/v1",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/school/attendance/mark": {
      "post": {
        "summary": "Mark attendance for a class",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "classId": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "attendance": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "studentId": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PRESENT",
                            "ABSENT",
                            "LATE",
                            "HALF_DAY",
                            "ON_LEAVE"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance marked"
          }
        }
      }
    },
    "/school/attendance/daily": {
      "get": {
        "summary": "Get daily attendance for a class",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily attendance"
          }
        }
      }
    },
    "/school/attendance/student/{studentId}/monthly": {
      "get": {
        "summary": "Get monthly attendance for a student",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student monthly attendance"
          }
        }
      }
    },
    "/school/attendance/class/monthly": {
      "get": {
        "summary": "Get class monthly attendance",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class monthly attendance"
          }
        }
      }
    },
    "/school/attendance/defaulters": {
      "get": {
        "summary": "Get attendance defaulters",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "threshold",
            "schema": {
              "type": "integer",
              "default": 75
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of defaulters"
          }
        }
      }
    },
    "/school/attendance/leave/request": {
      "post": {
        "summary": "Submit leave request",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "type": "string"
                  },
                  "fromDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "toDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leave request submitted"
          }
        }
      }
    },
    "/school/attendance/leave/{id}/status": {
      "put": {
        "summary": "Approve/Reject leave request",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "APPROVED",
                      "REJECTED"
                    ]
                  },
                  "remarks": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leave request updated"
          }
        }
      }
    },
    "/school/attendance/leave": {
      "get": {
        "summary": "Get all leave requests",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of leave requests"
          }
        }
      }
    },
    "/school/attendance/report": {
      "get": {
        "summary": "Get attendance report",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fromDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "pdf",
                "excel"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance report"
          }
        }
      }
    },
    "/school/attendance/notifications/send": {
      "post": {
        "summary": "Send attendance notifications",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "classId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notifications sent"
          }
        }
      }
    },
    "/school/attendance/stats": {
      "get": {
        "summary": "Get attendance statistics",
        "tags": [
          "School - Attendance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance statistics"
          }
        }
      }
    },
...
    "/school/stats": {
      "get": {
        "summary": "Get all stats",
        "tags": [
          "School - Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}